feat(study): Learn about IoC 23-10-04
Today, Learn about Inversion of Control. MemberService is object
that
have a behavior "welcome". To use "MemberService" object, we have
to
construct "MemberService" object by ourselves.(FC39.java)
But we make
AppConfig Class and we can pass Cotrolling the life cycle of an object
to AppConfig Class.
This is an indirect inversion of control.